Always print destructor errors
These errors are serious. However, per the C++ specification, they
can not be handled and logged via try catch blocks. Otherwise, the
program terminates. Although it is possible to use try catch block
in the destructor itself, it is much more straightforward to
unconditionally print out errors as an interim measure.