dht: add a transaction queue to avoid ID exhaustion
Use-after-free occurs when the transaction ID is exhausted.

As the ID only has one char, on an instance with many torrents, it
is easy to exhaust the ID space, so that new transactions have to
be discarded.

However, The new transaction could be associated with an ongoing
DhtSearch that does not expect transaction to be freed.

Follow up of "dht_server: fix potential use-after-free" (1418bf99),
properly fix the issue and deal with memory leak with this method.

Bug: rakshasa/libtorrent#166, rakshasa/libtorrent#209, rakshasa/libtorrent#944